pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "undertone"
version = "0.3.0"
description = "Voice typing for Linux. Talk it. Type it. Ship it."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "oneknight" }
]
keywords = ["voice", "typing", "dictation", "speech-to-text", "whisper", "linux"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Text Processing",
]
dependencies = [
    "faster-whisper>=1.1.0",
    "sounddevice>=0.5.0",
    "numpy>=1.24.0",
    "pynput>=1.7.0",
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "mypy>=1.8",
    "ruff>=0.3",
]

[project.urls]
Homepage = "https://github.com/oneKn8/undertone"
Repository = "https://github.com/oneKn8/undertone"
Issues = "https://github.com/oneKn8/undertone/issues"

[project.scripts]
undertone = "undertone.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/undertone"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

